⚡️ chore: Update npm dependencies and add postcss.config.cjs
This commit updates the npm dependencies by adding the "cssnano" package to the devDependencies in the package.json file. Additionally, it introduces a new file, postcss.config.cjs, which exports a PostCSS configuration object. The configuration includes the "cssnano" plugin when the NODE_ENV environment variable is set to "production". This change aims to optimize the CSS output in the production build.